[DP-32] - Перенес коллекцию из ReusableLikedMoviesUICollectionView в CoincidencesViewController. Удалил переиспользуемый класс ReusableLikedMoviesUICollectionView за ненадобностью.
